+++ src/storage-preview/azext_storage_preview/tests/latest/__init__.py
@@ -6,5 +6,5 @@
 from azure.cli.core.profiles import register_resource_type
 from ...profiles import CUSTOM_MGMT_STORAGE, CUSTOM_DATA_STORAGE_FILEDATALAKE
 
-register_resource_type('latest', CUSTOM_MGMT_STORAGE, '2021-09-01')
+register_resource_type('latest', CUSTOM_MGMT_STORAGE, '2022-09-01')
 register_resource_type('latest', CUSTOM_DATA_STORAGE_FILEDATALAKE, '2020-06-12')